Use float for amount. Commons methods not working

+++ transaction-service/src/main/java/tiktok/hackathon/model/TransactionFactory.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Component;
 public class TransactionFactory {
   public Transaction generate(
       String cardId,
-      int amount,
+      float amount,
       LocalDateTime transactionDateTime,
       String category,
       float lat,
